+++ src/udp/udp_handlers.js
@@ -11,47 +11,65 @@ module.exports.udpInitTaskErrorServer = function () {
     globals.udpServerTaskFailureSocket.on('listening', function (message, remote) {
         var address = globals.udpServerTaskFailureSocket.address();
 
-        globals.logger.info(`UDP server listening on ${address.address}:${address.port}`);
+        globals.logger.info(`TASKFAILURE: UDP server listening on ${address.address}:${address.port}`);
 
         // Publish MQTT message that UDP server has started
-        globals.mqttClient.publish(
-            globals.config.get('Butler.mqttConfig.taskFailureServerStatusTopic'),
-            'start',
-        );
+        globals.mqttClient.publish(globals.config.get('Butler.mqttConfig.taskFailureServerStatusTopic'), 'start');
     });
 
     // Handler for UDP error event
     globals.udpServerTaskFailureSocket.on('error', function (message, remote) {
         var address = globals.udpServerTaskFailureSocket.address();
-        globals.logger.error(`UDP server error on ${address.address}:${address.port}`);
+        globals.logger.error(`TASKFAILURE: UDP server error on ${address.address}:${address.port}`);
 
         // Publish MQTT message that UDP server has reported an error
-        globals.mqttClient.publish(
-            globals.config.get('Butler.mqttConfig.taskFailureServerStatusTopic'),
-            'error',
-        );
+        globals.mqttClient.publish(globals.config.get('Butler.mqttConfig.taskFailureServerStatusTopic'), 'error');
     });
 
     // Main handler for UDP messages relating to failed tasks
-    globals.udpServerTaskFailureSocket.on('message', function (message, remote) {
-        var msg = message.toString().split(';');
-        globals.logger.warning(
-            `${msg[0]}: Task "${msg[1]}" failed, associated with app "${msg[2]}"`,
-        );
+    globals.udpServerTaskFailureSocket.on('message', async function (message, remote) {
+        try {
+            var msg = message.toString().split(';');
+            globals.logger.warn(`TASKFAILURE: ${msg[0]}: Task ${msg[1]} failed, associated with app ${msg[2]}`);
 
-        // Post to Slack when a task has failed
-        globals.slackObj.send({
-            text: 'Failed task: "' + msg[1] + '", linked to app "' + msg[2] + '".',
-            channel: globals.config.get('Butler.slackConfig.taskFailureChannel'),
-            username: msg[0],
-            icon_emoji: ':ghost:',
-        });
+            // Post to Slack when a task has failed, if Slack is enabled
+            if (globals.config.has('Butler.slackConfig.enable') && globals.config.get('Butler.slackConfig.enable') == true) {
+                globals.slackObj.send({
+                    text: 'Failed task: "' + msg[1] + '", linked to app "' + msg[2] + '".',
+                    channel: globals.config.get('Butler.slackConfig.taskFailureChannel'),
+                    username: msg[0],
+                    icon_emoji: ':ghost:',
+                });
+            }
 
-        // Publish MQTT message when a task has failed
-        globals.mqttClient.publish(
-            globals.config.get('Butler.mqttConfig.taskFailureTopic'),
-            msg[1],
-        );
+            // Post to MS Teams when a task has failed, if Teams is enabled
+            if (globals.config.has('Butler.teamsConfig.enable') && globals.config.get('Butler.teamsConfig.enable') == true) {
+                await globals.teamsTaskFailureObj.send(
+                    JSON.stringify({
+                        '@type': 'MessageCard',
+                        '@context': 'https://schema.org/extensions',
+                        summary: 'A reload task has failed in Qlik Sense',
+                        themeColor: '0078D7',
+                        title: `Failed task: ${msg[1]}, app name "${msg[2]}"`,
+                        sections: [
+                            {
+                                // activityTitle: msg[2],
+                                // activitySubtitle: msg[1],
+
+                                text: 'Please refer to the QMC for further details',
+                            },
+                        ],
+                    }),
+                );
+            }
+
+            // Publish MQTT message when a task has failed, if MQTT enabled
+            if (globals.config.has('Butler.mqttConfig.enable') && globals.config.get('Butler.mqttConfig.enable') == true) {
+                globals.mqttClient.publish(globals.config.get('Butler.mqttConfig.taskFailureTopic'), msg[1]);
+            }
+        } catch (err) {
+            globals.logger.error('TASKFAILURE: Failed processing failed task event');
+        }
     });
 };
 
@@ -67,10 +85,7 @@ module.exports.udpInitSessionConnectionServer = function () {
         globals.logger.info(`UDP server listening on ${address.address}:${address.port}`);
 
         // Publish MQTT message that UDP server has started
-        globals.mqttClient.publish(
-            globals.config.get('Butler.mqttConfig.sessionServerStatusTopic'),
-            'start',
-        );
+        globals.mqttClient.publish(globals.config.get('Butler.mqttConfig.sessionServerStatusTopic'), 'start');
     });
 
     // Handler for UDP error event
@@ -80,10 +95,7 @@ module.exports.udpInitSessionConnectionServer = function () {
         globals.logger.error(`UDP server error on ${address.address}:${address.port}`);
 
         // Publish MQTT message that UDP server has reported an error
-        globals.mqttClient.publish(
-            globals.config.get('Butler.mqttConfig.sessionServerStatusTopic'),
-            'error',
-        );
+        globals.mqttClient.publish(globals.config.get('Butler.mqttConfig.sessionServerStatusTopic'), 'error');
     });
 
     // Main handler for UDP messages relating to session and connection events
@@ -101,32 +113,20 @@ module.exports.udpInitSessionConnectionServer = function () {
 
         // Handle session events
         if (msg[1] == 'Start session') {
-            globals.mqttClient.publish(
-                globals.config.get('Butler.mqttConfig.sessionStartTopic'),
-                msg[0] + ': ' + msg[2] + '/' + msg[3],
-            );
+            globals.mqttClient.publish(globals.config.get('Butler.mqttConfig.sessionStartTopic'), msg[0] + ': ' + msg[2] + '/' + msg[3]);
         }
 
         if (msg[1] == 'Stop session') {
-            globals.mqttClient.publish(
-                globals.config.get('Butler.mqttConfig.sessionStopTopic'),
-                msg[0] + ': ' + msg[2] + '/' + msg[3],
-            );
+            globals.mqttClient.publish(globals.config.get('Butler.mqttConfig.sessionStopTopic'), msg[0] + ': ' + msg[2] + '/' + msg[3]);
         }
 
         // Handle connection events
         if (msg[1] == 'Open connection') {
-            globals.mqttClient.publish(
-                globals.config.get('Butler.mqttConfig.connectionOpenTopic'),
-                msg[0] + ': ' + msg[2] + '/' + msg[3],
-            );
+            globals.mqttClient.publish(globals.config.get('Butler.mqttConfig.connectionOpenTopic'), msg[0] + ': ' + msg[2] + '/' + msg[3]);
         }
 
         if (msg[1] == 'Close connection') {
-            globals.mqttClient.publish(
-                globals.config.get('Butler.mqttConfig.connectionCloseTopic'),
-                msg[0] + ': ' + msg[2] + '/' + msg[3],
-            );
+            globals.mqttClient.publish(globals.config.get('Butler.mqttConfig.connectionCloseTopic'), msg[0] + ': ' + msg[2] + '/' + msg[3]);
         }
     });
 };