+++ changelog.md
@@ -6,10 +6,15 @@
 - Key-value store
 - Swagger API docs
 - Uptime logging, incl memory usage stored in Influx db for charting/monitoring in Grafana
+- Configurable Docker healthcheck
+- Added option to send task failure notifications to MS Teams
 - Updated dependencies (adresses security issues etc)
 - Breaking
   - http method changed for some API endpoints, with this follows changes in how parameters are passed (e.g. body parameters are used for PUT requests, rather than the query parameters that were previously used)
   - API endpoints are all lowercase
+  - Some keys in the main configuration have new names. For example, there was not a single way of using 'enable' vs 'enabled' keys.
+    The config file had both _Butler.heartbeat.enabled_, as well as _Butler.mqttConfig.enable_. 
+    Confusing - in v4 only _.enable_ is used.
 
 ## 3.1.0